Corrigir o produtor RabbitMQ e o estouro do numero de carro
Um canal AMQP por faixa no produtor, espelhando o consumidor. Canais
serializam operacoes, entao o canal unico dava ao produtor um quarto do
paralelismo do consumidor. O teto do RabbitMQ subiu de 27,7 mil para 42,7
mil ev/s -- o valor anterior era limite do instrumento, nao do broker. A
barreira de confirmacao caiu de 1.000 para 250 por faixa, ja que agora sao
P canais em voo simultaneo.

Estouro de tipo no numero do carro: a multiplicacao de frota soma
replica*100 ao numero original, e com fator 671 (necessario para 50 mil
ev/s) o valor passava de 32.767 e o short virava negativo, quebrando o
calculo da faixa. O campo passou a int, o codec de 44 para 46 bytes e as
colunas do banco de SMALLINT para INTEGER. Qualquer carga acima de ~30
mil ev/s estava afetada.

O jitter do produtor entra no criterio de saturacao do runner: sem ele o
script classificava como "ok" rodadas em que o gerador nao sustentou a
taxa -- que o proprio projeto define como invalidas.

Digest preservado apos a mudanca de codec: 4F4C77C2B84FD4A6.

+++ infra/postgres/init/01-schema.sql
@@ -8,7 +8,7 @@
 CREATE UNLOGGED TABLE processed_event (
     run_id          UUID        NOT NULL,
     architecture    TEXT        NOT NULL,   -- ex.: kafka-channels
-    driver_number   SMALLINT    NOT NULL,
+    driver_number   INTEGER     NOT NULL,
     session_key     INTEGER     NOT NULL,
     event_time      TIMESTAMPTZ NOT NULL,   -- timestamp original da telemetria
     published_at    TIMESTAMPTZ NOT NULL,   -- momento da publicacao no broker
@@ -25,7 +25,7 @@ CREATE UNLOGGED TABLE processed_event (
 CREATE UNLOGGED TABLE driver_window_stats (
     run_id          UUID        NOT NULL,
     architecture    TEXT        NOT NULL,
-    driver_number   SMALLINT    NOT NULL,
+    driver_number   INTEGER     NOT NULL,
     window_start    TIMESTAMPTZ NOT NULL,
     window_end      TIMESTAMPTZ NOT NULL,
     event_count     INTEGER     NOT NULL,
